Wrote tests for operator+= cstring.
Also updated the makefiles, they should work a little better now in
terms of where they put things. Due to us having our own cstddef, I had
to make the .h files be preprocessed before they are included in the
tests files -- shouldn't be much of a problem, but is ugly.